+++ app/routes/_layout+/settings.workspace.index.tsx
@@ -7,9 +7,9 @@ import ContextualModal from "~/components/layout/contextual-modal";
 import { ListHeader } from "~/components/list/list-header";
 import { ListItem } from "~/components/list/list-item";
 import { Badge } from "~/components/shared";
+import { ControlledActionButton } from "~/components/shared/controlled-action-button";
 import { Image } from "~/components/shared/image";
 import { UserBadge } from "~/components/shared/user-badge";
-import { ControlledActionButton } from "~/components/subscription/controlled-action-button";
 import { Table, Td, Th } from "~/components/table";
 import { WorkspaceActionsDropdown } from "~/components/workspace/workspace-actions-dropdown";
 import { db } from "~/database";
